src/robot_state_machine/robot_state_machine/utils/robot-video-stream/requirements.txt
pyzmq>=25.0.0
opencv-python>=4.8.0
numpy>=1.24.0
loguru>=0.7.0
pyyaml>=6.0.0
